Update README.md
The selected dataset provided in this project is gathered from a telecommunications company in line with its customer retention program. The focus of the program is to examine the total number of existing and churned customers of the company. The resultant information will further be used to predict the number of customers that are most likely to churn. The company is therefore interested in building a classification model for churn prediction, which will help to achieve the objective of minimizing the churn rate and consequently enhance customer retention. Identifying the factors that would likely cause a customer to churn, will therefore be of key importance.